requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.4
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
bcrypt==4.3.0
beautifulsoup4==4.12.3
billiard==4.2.1
bleach==6.1.0
blinker==1.9.0
bootstrap4==0.1.0
cachetools==5.5.2
captcha==0.7.1
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
crispy-bootstrap4==2022.1
cryptography==45.0.3
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.18
dj-database-url==2.3.0
Django==5.0.14
django-activity-stream==2.0.0
django-authtools==2.0.1
django-booking==0.7.2
django-bootstrap==0.2.4
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap4==25.1
django-braces==1.17.0
django-comments-dab==2.8.0
django-common==0.1.51
django-countries==7.6.1
django-crispy-forms==2.0
django-etc==1.4.0
django-fontawesome==1.0
django-forms-bootstrap==3.1.0
django-heroku==0.3.1
django-hitcount==1.3.5
django-hosts==6.0
django-hvad==1.8.0
django-jsonfield==1.4.1
django-libs==2.0.3
django-nose==1.4.7
django-paypal==2.1
django-phone-verify==3.1.0
django-phonenumber-field==8.1.0
django-realtime==1.1
django-reservation==0.2.10
django-rest-framework==0.1.0
django-social-auth==0.7.28
django-social-share==2.3.0
django-static-fontawesome==6.5.2.0
django-tempus-dominus==5.1.2.17
django-ticketing==0.7.4
django-uuslug==2.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
et-xmlfile==1.1.0
executing==2.1.0
extra-streamlit-components==0.1.80
fastjsonschema==2.20.0
Flask==3.0.3
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.6.0
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
json5==0.9.25
jsonfield==3.1.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kombu==5.5.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
multidict==6.4.4
mysql-connector-python==9.3.0
narwhals==1.41.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nexmo==2.5.2
nose==1.3.7
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.2.6
oauth2==1.9.0.post1
oauthlib==3.2.2
ollama==0.4.6
openpyxl==3.1.5
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
phonenumbers==9.0.6
phonenumberslite==9.0.6
pillow==11.2.1
plotly==5.24.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.3.1
protobuf==6.31.1
psutil==6.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure_eval==0.2.3
py4j==0.10.9.9
pyarrow==20.0.0
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.2.0
pyspark==4.0.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-json-logger==2.0.7
python-openid==2.2.5
python-slugify==8.0.4
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2025.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
Send2Trash==1.8.3
setuptools==75.1.0
simplejson==3.20.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sorl-thumbnail==12.11.0
soupsieve==2.6
South==1.0.2
stack-data==0.6.3
streamlit==1.45.1
streamlit-authenticator==0.4.2
tenacity==9.0.0
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.3.0
toml==0.10.2
tornado==6.4.1
traitlets==5.14.3
twilio==9.6.2
types-python-dateutil==2.9.0.20240821
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
urllib3==2.4.0
uv==0.2.35
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
wheel==0.44.0
whitenoise==6.5.0
widgetsnbextension==4.0.13
wrapt==1.17.2
yarl==1.20.0
zipp==3.21.0
